refactor: change namespace of Parameters
