+++ src/CuiLib/Commands/Command.cs
@@ -3,6 +3,7 @@ using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using CuiLib.Options;
+using CuiLib.Parameters;
 
 namespace CuiLib.Commands
 {
