+++ src/CuiLib/Parameters/Parameter.cs
@@ -3,7 +3,7 @@ using System.Diagnostics.CodeAnalysis;
 using CuiLib.Checkers;
 using CuiLib.Converters;
 
-namespace CuiLib.Options
+namespace CuiLib.Parameters
 {
     /// <summary>
     /// パラメータ引数を表します。
